# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fruit & Dry Fruits Shop Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: featured
          in: query
          schema:
            type: boolean
        - name: organic
          in: query
          schema:
            type: boolean
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, price, rating, newest, bestseller]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"

    post:
      operationId: createProduct
      tags: [products, admin]
      summary: Create a product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products, admin]
      summary: Update a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products, admin]
      summary: Delete a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user's orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── REVIEWS ───────────────────────────────────────────────────────────
...
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin, orders]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin, orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── ADMIN: CUSTOMERS ─────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Product:
      type: object
      required: [id, name, slug, categoryId, categoryName, price, discountPrice, stock, imageUrl, rating, reviewCount, isFeatured, isOrganic, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        stock:
          type: integer
        imageUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        isOrganic:
          type: boolean
        weightOptions:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProductDetail:
      type: object
      required: [id, name, slug, categoryId, categoryName, price, discountPrice, stock, imageUrl, rating, reviewCount, isFeatured, isOrganic, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        stock:
          type: integer
        imageUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        isOrganic:
          type: boolean
        weightOptions:
          type: ["string", "null"]
        nutritionInfo:
          type: ["string", "null"]
        healthBenefits:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    ProductList:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProductInput:
      type: object
      required: [name, categoryId, price, stock]
      properties:
        name:
          type: string
        categoryId:
          type: integer
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        stock:
          type: integer
        imageUrl:
          type: string
        isFeatured:
          type: boolean
        isOrganic:
          type: boolean
        weightOptions:
          type: string
        nutritionInfo:
          type: string
        healthBenefits:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        stock:
          type: integer
        imageUrl:
          type: string
        isFeatured:
          type: boolean
        isOrganic:
          type: boolean
        weightOptions:
          type: string
        nutritionInfo:
          type: string
        healthBenefits:
          type: string

    # ── CART ──
...
    OrderItem:
      type: object
      required: [id, productId, productName, productImage, quantity, price, weightOption, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
        weightOption:
          type: ["string", "null"]
        subtotal:
          type: number
    Order:
      type: object
      required: [id, userId, status, deliveryAddress, phone, paymentMethod, subtotal, discount, deliveryCharge, total, items, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, packed, shipped, out_for_delivery, delivered, cancelled]
        deliveryAddress:
          type: string
        phone:
          type: string
        paymentMethod:
          type: string
          enum: [cash_on_delivery, upi, debit_card, credit_card]
        subtotal:
          type: number
        discount:
          type: number
        deliveryCharge:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
    OrderInput:
      type: object
      required: [deliveryAddress, phone, paymentMethod]
      properties:
        deliveryAddress:
          type: string
        phone:
          type: string
        paymentMethod:
          type: string
          enum: [cash_on_delivery, upi, debit_card, credit_card]
        couponCode:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, packed, shipped, out_for_delivery, delivered, cancelled]

    # ── REVIEWS ──
    Review:
      type: object
      required: [id, productId, userId, userName, rating, comment, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
...