# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Fruit & Dry Fruits Shop Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  status: OrderStatus;
  deliveryAddress: string;
  phone: string;
  paymentMethod: OrderPaymentMethod;
  subtotal: number;
  discount: number;
  deliveryCharge: number;
  total: number;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  estimatedDelivery?: string | null;
  items: OrderItem[];
  createdAt: Date;
}