# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Fruit & Dry Fruits Shop Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  categoryId: number;
  description?: string;
  price: number;
  discountPrice?: number;
  stock: number;
  imageUrl?: string;
  isFeatured?: boolean;
  isOrganic?: boolean;
  weightOptions?: string;
  nutritionInfo?: string;
  healthBenefits?: string;
}